--- examples/daymet-hi-annual-template.json
@@ -1,44 +1,0 @@
-{
-  "stac_version": "1.0.0",
-  "type": "Collection",
-  "id": "daymet-hi-annual",
-  "title": "Daymet Hawaii Annual",
-  "description": "Daymet is a data product derived from a collection of algorithms and computer software designed to interpolate and extrapolate from daily meteorological observations to produce gridded estimates of daily weather parameters. Weather parameters generated include daily surfaces of minimum and maximum temperature, precipitation, vapor pressure, radiation, snow water equivalent, and day length produced on a 1 km x 1 km gridded surface. The motivation for producing Daymet is to provide measurements of near-surface meteorological conditions where no instrumentation exists. Having estimates of these surfaces is critical to understanding many processes in the terrestrial biogeochemical system.",
-  "license": "proprietary",
-  "extent": {
-    "spatial": {
-      "bbox": [
-        [
-          -160.3056,
-          17.9539,
-          -154.772,
-          23.5186
-        ]
-      ]
-    },
-    "temporal": {
-      "interval": [
-        [
-          "1980-01-01T12:00:00Z",
-          "2019-12-31T12:00:00Z"
-        ]
-      ]
-    }
-  },
-  "links": [],
-  "assets": {
-    "data": {
-      "href": "abfs://daymet-zarr/annual/hi.zarr",
-      "title": "Annual Hawaii Daymet zarr root",
-      "description": "The root of the annual Hawaii Daymet Zarr Group on Azure Blob Storage.",
-      "roles": [
-        "data",
-        "zarr-root",
-        "azure-blob-storge"
-      ],
-      "storage_options": {
-        "account_name": "daymeteuwest"
-      }
-    }
-  }
-}
