+++ test_xstac.py
@@ -3,6 +3,7 @@ import xarray as xr
 import numpy as np
 import pandas as pd
 import pytest
+import pystac
 
 
 data = np.empty((40, 584, 284), dtype="float32")
@@ -135,4 +136,7 @@ def test_xarray_to_stac(ds):
         x_dimension="x",
         y_dimension="y",
     )
-    assert result is not None
+    assert result.id == "id"
+    assert isinstance(result, pystac.Collection)
+    assert result.description == "description"
+    assert result.license == "license"
